translations-api/requirements.txt
flask-restplus==0.12.1
flask-cors==3.0.7

